requirements.txt
# --- Core FastAPI ---
fastapi==0.115.0
uvicorn==0.30.1
python-multipart==0.0.9
requests==2.32.3
httpx==0.27.2

# --- Data Handling ---
pandas==2.2.3
openpyxl==3.1.5
numpy==1.26.4

# --- PDF/Text Processing ---
pypdf==4.3.1
beautifulsoup4==4.12.3
nltk==3.9.1

# --- Embeddings and Vector DB ---
sentence-transformers==3.0.1
scikit-learn==1.5.2
qdrant-client==1.10.1

# --- Audio and Speech (TTS/STT) ---
sounddevice==0.4.7
pydub==0.25.1
SpeechRecognition==3.10.4
gTTS==2.5.3
torch==2.4.1
torchaudio==2.4.1

# --- Utilities ---
tqdm==4.66.5
python-dotenv==1.0.1
jinja2==3.1.4

openai==2.7.2

starlette==0.38.6






